+++ demisto_sdk/commands/common/constants.py
@@ -1002,3 +1002,6 @@ MAX_FETCH = 'max_fetch'
 OLDEST_SUPPORTED_VERSION = '5.0.0'
 
 FEATURE_BRANCHES = ['v4.5.0']
+
+SKIP_RELEASE_NOTES_FOR_TYPES = (FileType.RELEASE_NOTES, FileType.README, FileType.TEST_PLAYBOOK,
+                                FileType.TEST_SCRIPT, FileType.IMAGE, FileType.DOC_IMAGE)