+++ demisto_sdk/commands/validate/tests/validators_test.py
@@ -1180,3 +1180,21 @@ def test_should_raise_pack_version(pack_name, expected):
     validate_manager = ValidateManager()
     res = validate_manager.should_raise_pack_version(pack_name)
     assert res == expected
+
+
+def test_run_validation_using_git_on_only_metadata_changed(mocker):
+    """
+    Given
+        - metadata file that was changed.
+    When
+        - Run all tests on the file.
+    Then
+        - validate That no error returns.
+    """
+    mocker.patch.object(ValidateManager, 'setup_git_params')
+    mocker.patch.object(ValidateManager, 'get_modified_and_added_files',
+                        return_value=(set(), set(), set(), {'Packs/TestPack/pack_metadata.json'}, {'TestPack'}))
+
+    validate_manager = ValidateManager()
+    res = validate_manager.run_validation_using_git()
+    assert res