+++ CHANGELOG.md
@@ -12,6 +12,8 @@
 * Fixed an issue in the **lint** command where *finally* clause was not supported in main function.
 * Fixed an issue in the **validate** command where changing any entity ID was not validated.
 * Fixed an issue in the **validate** command where *--staged* flag did not bring only changed files.
+* Fixed the **update-release-notes** command to ignore changes in the metadata file.
+* Fixed the **validate** command to ignore metadata changes when checking if a version bump is needed.
 
 
 # v1.2.8